refactor - Clean up kts implementation
