--- kotlin-asyncapi-maven-plugin/src/main/kotlin/org/openfolder/kotlinasyncapi/mavenplugin/AsyncApiScriptHost.kt
@@ -1,30 +1,0 @@
-package org.openfolder.kotlinasyncapi.mavenplugin
-
-import org.openfolder.kotlinasyncapi.model.AsyncApi
-import org.openfolder.kotlinasyncapi.script.AsyncApiScript
-import java.io.File
-import javax.inject.Named
-import javax.inject.Singleton
-import kotlin.script.experimental.api.implicitReceivers
-import kotlin.script.experimental.host.toScriptSource
-import kotlin.script.experimental.jvmhost.BasicJvmScriptingHost
-
-@Singleton
-@Named
-internal class AsyncApiScriptHost : AsyncApiScriptRunner {
-
-    private val jvmScriptingHost = BasicJvmScriptingHost()
-
-    override fun run(script: File): AsyncApi {
-        val scriptReceiver = AsyncApi()
-
-        jvmScriptingHost.evalWithTemplate<AsyncApiScript>(
-            script = script.toScriptSource(),
-            evaluation = {
-                implicitReceivers(scriptReceiver)
-            }
-        )
-
-        return scriptReceiver
-    }
-}
